LayoutTests/imported/w3c/web-platform-tests/css/selectors/parsing/parse-has-disallow-nesting-has-inside-has-expected.txt


PASS ".a:has(.b:has(.c))" should be an invalid selector
FAIL ".a:has(:is(.b:has(.c)))" should be a valid selector assert_equals: CSS.supports() reports the expected value expected true but got false
FAIL ".a:has(:is(.b:has(.c), .d))" should be a valid selector assert_equals: CSS.supports() reports the expected value expected true but got false
